split tests properly
