+++ src/Framework/vite-bt-plugin/hash.test.ts
@@ -1,0 +1,10 @@
+import { expect, it } from 'vitest';
+import { readTsx, transformCode } from './test.helper.ts';
+
+it('calculates persistent hash based on test source code, excluding extra indentation', async () => {
+  const fixturePath = './fixtures/hash.fixture.tsx';
+  const input = await readTsx(fixturePath);
+  const output = await transformCode(input);
+
+  expect(output).toMatchSnapshot();
+});
